Use Long arithmetic for the plane bounds check
Review follow-up on PR #14: the single up-front bounds check could
overflow Int for pathological stride/offset combinations and silently
pass; computing it in Long keeps the descriptive IllegalArgumentException
for every geometry.

+++ app/src/main/java/ch/lkmc/kirsch/capture/Yuv420Packer.kt
@@ -83,7 +83,11 @@ object Yuv420Packer {
         val base = buffer.position()
         // Strides are positive, so the last pixel of the last row is the
         // highest index ever read; checking it once bounds every access.
-        val lastIndex = base + (startY + height - 1) * rowStride + (startX + width - 1) * pixelStride
+        // Long arithmetic keeps the check valid even for geometries whose
+        // byte offsets would overflow Int.
+        val lastIndex = base.toLong() +
+            (startY + height - 1).toLong() * rowStride +
+            (startX + width - 1).toLong() * pixelStride
         require(lastIndex < buffer.limit()) {
             "plane buffer is too small for declared strides and crop"
         }
